backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "bcrypt>=4.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.32",
    "uvicorn>=0.40.0",
    "requests>=2.32.0",
    "pyjwt>=2.10.0",
]
